Add task lists and a drawer storing them

+++ app/src/main/java/com/xenon/todolist/TaskList.kt
@@ -8,7 +8,7 @@ data class TaskList(
     override var id: Int,
     var name: String,
     var items: ArrayList<TaskItem>,
-    var createdTime: Int
+    var createdTime: Long
 ) : LiveListItem {
 
 }
